+++ src/Symfony/Component/EventDispatcher/Tests/DependencyInjection/RegisterListenersPassTest.php
@@ -660,6 +660,242 @@ class RegisterListenersPassTest extends TestCase
         $this->assertSame(['decorator'], $listeners);
     }
 
+    public function testListenersWithoutConstraintsAreRegisteredInTagOrder()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'priority' => 10]);
+        $container->register('c', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $expectedCalls = [
+            ['addListener', ['foo', [new ServiceClosureArgument(new Reference('a')), 'onEvent'], 0]],
+            ['addListener', ['foo', [new ServiceClosureArgument(new Reference('b')), 'onEvent'], 10]],
+            ['addListener', ['foo', [new ServiceClosureArgument(new Reference('c')), 'onEvent'], 0]],
+        ];
+
+        $this->assertEquals($expectedCalls, $container->getDefinition('event_dispatcher')->getMethodCalls());
+    }
+
+    public function testABeforeConstraintReordersListenersOfTheSamePriority()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('c', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'before' => 'a']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'c', 0], ['foo', 'a', 0], ['foo', 'b', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAListenerCanBeInsertedBetweenTwoOthers()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('x', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'bar', 'method' => 'onEvent']);
+        $container->register('c', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('e', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'after' => 'b', 'before' => 'c']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'b', 0], ['bar', 'x', 0], ['foo', 'e', 0], ['foo', 'c', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAConstraintRaisesThePriorityOfTheConstrainedListener()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'priority' => 10]);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'before' => 'a']);
+        $container->register('d', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'priority' => 5]);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'b', 10], ['foo', 'a', 10], ['foo', 'd', 5]], $this->getListenerCalls($container));
+    }
+
+    public function testConstraintsTargetingAnUnknownListenerAreIgnored()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'after' => 'from_a_bundle_that_is_not_installed']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'a', 0], ['foo', 'b', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAListenerCanBeOrderedAfterASubscriber()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'after' => 'subscriber']);
+        $container->register('subscriber', SubscriberService::class)->addTag('kernel.event_subscriber');
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'subscriber', 0], ['event', 'a', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAConstraintCanTargetAClass()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'after' => SubscriberService::class]);
+        $container->register('subscriber', SubscriberService::class)->addTag('kernel.event_subscriber');
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'subscriber', 0], ['event', 'a', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testSeveralRegistrationsOfTheSameServiceKeepTheirOwnPlace()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'priority' => -200, 'before' => 'b']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent']);
+        $container->register('subscriber', MultiPrioritySubscriber::class)->addTag('kernel.event_subscriber');
+
+        (new RegisterListenersPass())->process($container);
+
+        $expectedCalls = [['event', 'subscriber', 100], ['event', 'a', 0], ['event', 'b', 0], ['event', 'subscriber', -100]];
+
+        $this->assertSame($expectedCalls, $this->getListenerCalls($container));
+    }
+
+    public function testConstraintsAreScopedToTheirDispatcher()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('other_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'dispatcher' => 'other_dispatcher']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'before' => 'a']);
+        $container->register('c', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'b', 0], ['foo', 'c', 0]], $this->getListenerCalls($container));
+        $this->assertSame([['foo', 'a', 0]], $this->getListenerCalls($container, 'other_dispatcher'));
+    }
+
+    public function testCyclicConstraintsAreReported()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'before' => 'b']);
+        $container->register('b', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent', 'before' => 'a']);
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Invalid "before"/"after" constraints for event "foo": cycle detected in the "before"/"after" constraints: "a" -> "b" -> "a".');
+
+        (new RegisterListenersPass())->process($container);
+    }
+
+    public function testBeforeAndAfterAreReadFromTheAttribute()
+    {
+        $container = $this->createContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'foo', 'method' => 'onEvent']);
+        $container->register('b', OrderedListener::class)->setAutoconfigured(true);
+
+        (new AttributeAutoconfigurationPass())->process($container);
+        (new ResolveInstanceofConditionalsPass())->process($container);
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['foo', 'b', 0], ['foo', 'a', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAConstraintCanTargetASingleMethodOfAListener()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('sub', MultiPrioritySubscriber::class)->addTag('kernel.event_subscriber');
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'priority' => -200, 'before' => 'sub::onLate']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'sub', 100], ['event', 'a', -100], ['event', 'sub', -100]], $this->getListenerCalls($container));
+    }
+
+    public function testTargetingTheWholeServiceOutranksAllOfItsMethods()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('sub', MultiPrioritySubscriber::class)->addTag('kernel.event_subscriber');
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'before' => 'sub']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'a', 100], ['event', 'sub', 100], ['event', 'sub', -100]], $this->getListenerCalls($container));
+    }
+
+    public function testAMethodTargetCanUseTheClassName()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('sub', MultiPrioritySubscriber::class)->addTag('kernel.event_subscriber');
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'priority' => -200, 'before' => MultiPrioritySubscriber::class.'::onLate']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'sub', 100], ['event', 'a', -100], ['event', 'sub', -100]], $this->getListenerCalls($container));
+    }
+
+    public function testAMethodTargetOnAKnownListenerMustExist()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('sub', MultiPrioritySubscriber::class)->addTag('kernel.event_subscriber');
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'before' => 'sub::onLat']);
+
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Invalid "before" constraint on listener "a": "sub" does not listen to event "event" with method "onLat".');
+
+        (new RegisterListenersPass())->process($container);
+    }
+
+    public function testAMethodTargetOnAnAbsentServiceIsIgnored()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'before' => 'from_an_absent_bundle::onWhatever']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'a', 0]], $this->getListenerCalls($container));
+    }
+
+    public function testAMethodTargetResolvesAgainstASubscriberUsingNamedKeys()
+    {
+        $container = new ContainerBuilder();
+        $container->register('event_dispatcher', \stdClass::class);
+        $container->register('sub', NamedKeysSubscriber::class)->addTag('kernel.event_subscriber');
+        $container->register('a', InvokableListenerService::class)->addTag('kernel.event_listener', ['event' => 'event', 'method' => 'onEvent', 'priority' => -200, 'before' => 'sub::onLate']);
+
+        (new RegisterListenersPass())->process($container);
+
+        $this->assertSame([['event', 'sub', 100], ['event', 'a', -100], ['event', 'sub', -100]], $this->getListenerCalls($container));
+    }
+
+    private function getListenerCalls(ContainerBuilder $container, string $dispatcher = 'event_dispatcher'): array
+    {
+        $calls = [];
+        foreach ($container->getDefinition($dispatcher)->getMethodCalls() as [$method, $arguments]) {
+            if ('addListener' === $method) {
+                $calls[] = [$arguments[0], (string) $arguments[1][0]->getValues()[0], $arguments[2]];
+            }
+        }
+
+        return $calls;
+    }
+
     private function createContainerBuilder(): ContainerBuilder
     {
         $container = new ContainerBuilder();
@@ -774,6 +1010,49 @@ final class IncompleteSubscriber implements EventSubscriberInterface
     }
 
     public function __invoke(CustomEvent $event): void
+    {
+    }
+}
+
+#[AsEventListener(event: 'foo', method: 'onEvent', before: 'a')]
+final class OrderedListener
+{
+    public function onEvent(): void
+    {
+    }
+}
+
+final class MultiPrioritySubscriber implements EventSubscriberInterface
+{
+    public static function getSubscribedEvents(): array
+    {
+        return ['event' => [['onEarly', 100], ['onLate', -100]]];
+    }
+
+    public function onEarly(): void
+    {
+    }
+
+    public function onLate(): void
+    {
+    }
+}
+
+final class NamedKeysSubscriber implements EventSubscriberInterface
+{
+    public static function getSubscribedEvents(): array
+    {
+        return ['event' => [
+            ['method' => 'onEarly', 'priority' => 100],
+            ['method' => 'onLate', 'priority' => -100],
+        ]];
+    }
+
+    public function onEarly(): void
+    {
+    }
+
+    public function onLate(): void
     {
     }
 }